+++ projects/angular/src/forms/styles/_file-input.clarity.scss
@@ -69,13 +69,15 @@
     .clr-file-label {
       white-space: wrap;
       height: unset;
+      gap: tokens.$cds-global-space-5;
       margin-right: tokens.$cds-global-space-2;
+      padding-right: tokens.$cds-global-space-3;
     }
 
     .clr-file-clear-button {
-      padding-right: tokens.$cds-global-space-2;
-      padding-left: tokens.$cds-global-space-2;
       margin: 0;
+      padding-right: tokens.$cds-global-space-1;
+      padding-left: tokens.$cds-global-space-1;
       height: tokens.$cds-alias-typography-caption-line-height; // match label text line height
 
       cds-icon {
